Split components in frontend
