+++ feedback_frontend/src/components/mod.rs
@@ -1,0 +1,3 @@
+pub mod slider;
+pub mod cookie_banner;
+pub mod footer;
