+++ feedback_frontend/src/main.rs
@@ -1,9 +1,16 @@
+mod components;
+mod functions;
+
 use gloo::net::http::Request;
 use serde::Serialize;
 use wasm_bindgen::prelude::*;
 use web_sys::HtmlTextAreaElement;
 use yew::platform::spawn_local;
 use yew::prelude::*;
+use crate::components::cookie_banner::CookieBanner;
+use crate::components::footer::Footer;
+use crate::components::slider::Slider;
+use crate::functions::{format_thanks, input};
 
 const POST_URI: &str = include_str!("../target_uri.txt");
 const LORIS_LINK: &str = "https://www.youtube.com/channel/UCe40qwYch8JcmBST_BWaYNA";
@@ -48,195 +55,13 @@ fn app() -> Html {
     html! {
         <>
             <h1 class={classes!("text-3xl", "font-bold", "mb-6")}>{ "Feedback" }</h1>
-            { slider() }
+            <Slider/>
             { format_thanks(&thanks_message, &thanks_colour) }
             <p class={classes!("text-lg", "mb-4")}>{ "Please enter the Feedback here:" }</p>
             { input(&feedback, &thanks_message, &thanks_colour) }
-            <LorisFooter/>
+            <Footer/>
             <CookieBanner/>
         </>
     }
 }
 
-fn slider() -> Html {
-    html! {
-        <div class={classes!("absolute", "top-4", "right-4")}>
-            <label class={classes!("switch")}>
-                <input type="checkbox" id="theme-switch" onclick={Callback::from(|_| toggleTheme())} checked={true}/>
-                <span class={classes!("slider")}></span>
-            </label>
-        </div>
-    }
-}
-
-fn format_thanks(
-    thanks_message: &UseStateHandle<Option<String>>,
-    thanks_colour: &UseStateHandle<Colour>,
-) -> Html {
-    match &**thanks_message {
-        Some(msg) => html! {
-            <div class={classes!("w-full", "max-w-lg", "text-center", "mb-4")} style={format!("color: {}", thanks_colour.to_string())}>
-                { msg }
-            </div>
-        },
-        None => html! {},
-    }
-}
-
-fn input(
-    feedback: &UseStateHandle<String>,
-    thanks_msg: &UseStateHandle<Option<String>>,
-    thanks_colour: &UseStateHandle<Colour>,
-) -> Html {
-    let on_feedback_input = {
-        let feedback = feedback.clone();
-        Callback::from(move |e: InputEvent| {
-            if let Some(input) = e.target_dyn_into::<HtmlTextAreaElement>() {
-                feedback.set(input.value());
-            }
-        })
-    };
-
-    let on_click = {
-        let feedback = feedback.clone();
-        let thanks_msg = thanks_msg.clone();
-        let thanks_colour = thanks_colour.clone();
-
-        Callback::from(move |_| {
-            let feedback = feedback.clone();
-            let thanks_msg = thanks_msg.clone();
-            let thanks_colour = thanks_colour.clone();
-
-            if feedback.is_empty() {
-                thanks_colour.set(Colour::Red);
-                thanks_msg.set(Some(String::from("Please enter feedback!")));
-                return;
-            }
-
-            let feedback_data = Feedback { feedback: (*feedback).trim().to_string() };
-            let parsed_feedback = serde_json::to_string(&feedback_data).unwrap();
-
-            spawn_local(async move {
-                let response = Request::post(&POST_URI)
-                    .header("Content-Type", "application/json")
-                    .body(&parsed_feedback)
-                    .expect("Failed to create request")
-                    .send()
-                    .await;
-
-                match response {
-                    Ok(resp) if resp.ok() => {
-                        thanks_colour.set(Colour::Green);
-                        thanks_msg.set(Some(String::from("Thank you for your feedback!")));
-                        feedback.set(String::new());
-                    }
-                    Ok(resp) => {
-                        thanks_colour.set(Colour::Orange);
-                        thanks_msg.set(Some(format!("Backend was unable to handle request: \
-                            {resp:?}")))
-                    }
-                    Err(e) => {
-                        thanks_colour.set(Colour::Red);
-                        thanks_msg.set(Some(
-                            format!("Unable to send request: {e}.")
-                        ));
-                    }
-                }
-            });
-        })
-    };
-
-    html! {
-        <div class={classes!("w-full", "max-w-lg")}>
-            <textarea
-                id="textbox"
-                name="textbox"
-                rows="8"
-                class={classes!("w-full", "p-2", "bg-gray-100", "dark:bg-gray-800", "border", "border-gray-300", "dark:border-gray-700", "rounded", "mb-4", "resize-none")}
-                value={feedback.clone().to_string()}
-                oninput={on_feedback_input}
-            ></textarea>
-            <button
-                type="submit"
-                onclick={on_click}
-                class={classes!("w-full", "bg-indigo-700", "hover:bg-indigo-800", "text-white", "font-bold", "py-2", "px-4", "rounded")}
-            >
-                { "Submit" }
-            </button>
-        </div>
-    }
-}
-
-#[cfg(not(feature = "cookie_banner"))]
-#[function_component(CookieBanner)]
-fn cookie_banner() -> Html {
-    html! {}
-}
-
-#[cfg(feature = "cookie_banner")]
-#[function_component(CookieBanner)]
-fn cookie_banner() -> Html {
-    let show_banner = use_state(|| true);
-
-    if !*show_banner {
-        return html! {};
-    }
-
-    html! {
-        <div class={classes!(
-            "fixed",
-            "bottom-4",
-            "left-1/2",
-            "transform",
-            "-translate-x-1/2",
-            "bg-gray-100",
-            "dark:bg-gray-800",
-            "p-4",
-            "flex",
-            "justify-center",
-            "items-center",
-            "shadow-lg",
-            "rounded-lg",
-            "border",
-            "border-gray-200",
-            "dark:border-gray-700",
-            "max-w-xl",
-            "w-[95%]",
-        )}>
-            <div class={classes!("flex", "flex-col", "sm:flex-row", "items-center", "gap-4", "text-center")}>
-                <p class={classes!("text-gray-800", "dark:text-gray-200")}>
-                    { "We don't use any cookies." }
-                </p>
-                <button
-                    onclick={
-                        let show_banner = show_banner.clone();
-                        Callback::from(move |_| show_banner.set(false))
-                    }
-                    class={classes!(
-                        "px-4",
-                        "py-2",
-                        "bg-indigo-700",
-                        "hover:bg-indigo-800",
-                        "text-white",
-                        "rounded",
-                        "text-sm",
-                        "whitespace-nowrap"
-                    )}
-                >
-                    { "Got it, don't show again" }
-                </button>
-            </div>
-        </div>
-    }
-}
-
-#[function_component(LorisFooter)]
-fn loris_footer() -> Html {
-    html! {
-        <p class={classes!("text-sm", "italic", "mt-8")}>
-            { "Thank you, " }
-            <a href={LORIS_LINK} class={classes!("text-blue-500", "underline")} target="_blank">{ "Loris," }</a>
-            { " for the design" }
-        </p>
-    }
-}
